move EncryptedPayload to correct 'data' folder
